refactor(firmware): split FirmwareUpdateService into focused collaborators (part of #344)
FirmwareUpdateService had grown to 2,656 lines co-mingling two independent
flows behind one class: the PIC32 bootloader path and the WiFi-module path,
plus shared retry/state helpers and an embedded verify-progress parser.

Split behind the unchanged IFirmwareUpdateService facade:

- FirmwareUpdateContext   - shared state machine, progress, retry and
                            per-state timeout plumbing
- Pic32BootloaderSession  - the low-level HID bootloader exchanges
- Pic32FirmwareUpdater    - PIC32 flow orchestration + diagnostics + cleanup
- WifiModuleUpdater       - WINC version probe + external flash-tool run
- WifiFlashProgressParser - lifted out of its nested position

Pure refactor: no public API change and no behavior change. Verified by
diffing the normalized statement multiset of the original file against all
six new files - every difference is structural (signatures, fields, wiring),
with zero logic statements added, removed or altered.

Adds one test. The StateChanged sender was previously an implicit `this` on
a field-like event and so could not be wrong; it is now forwarded from the
context, which nothing asserted. The new test pins it (mutation-verified: it
fails when the sender is wrong).

WifiFlashProgressParser was internal nested in a public class, so its
external accessibility is unchanged; only the in-assembly name moved, which
is why three test references were updated.
